+++ apollo-compiler/src/main/kotlin/com/apollographql/apollo3/compiler/codegen/kotlin/KotlinCodeGen.kt
@@ -42,6 +42,7 @@ class KotlinCodeGen(
     private val packageNameGenerator: PackageNameGenerator,
     private val schemaPackageName: String,
     private val useSchemaPackageNameForFragments: Boolean,
+    private val typePackageName: String,
     /**
      * The operation id cannot be set in [IrOperation] because it needs access to [IrOperation.sourceWithFragments]
      * So we do this in the codegen step
@@ -76,7 +77,8 @@ class KotlinCodeGen(
         useSemanticNaming = useSemanticNaming,
         packageNameGenerator = packageNameGenerator,
         schemaPackageName = schemaPackageName,
-        useSchemaPackageNameForFragments = useSchemaPackageNameForFragments
+        useSchemaPackageNameForFragments = useSchemaPackageNameForFragments,
+        typePackageName = typePackageName,
     )
 
     val context = KotlinContext(