+++ apollo-gradle-plugin/src/main/kotlin/com/apollographql/apollo3/gradle/internal/DefaultApolloExtension.kt
@@ -532,6 +532,7 @@ abstract class DefaultApolloExtension(
       task.flattenModels.set(service.flattenModels)
       @OptIn(ApolloExperimental::class)
       task.generateTestBuilders.set(service.generateTestBuilders)
+      task.typePackageName.set(service.typePackageName)
       task.useSchemaPackageNameForFragments.set(service.useSchemaPackageNameForFragments)
       task.sealedClassesForEnumsMatching.set(service.sealedClassesForEnumsMatching)
       task.generateOptionalOperationVariables.set(service.generateOptionalOperationVariables)